Add files via upload
A Python program utilizing the Pandas package to compute and store (within a DataFrame) the monthly 5 year average prices for the popular cryptocurrency, Litecoin. From the monthly 5 year prices, statistics such as the minimum/maximum value, mean/median and 25th/75th percentile are computed and stored (within a separate DataFrame).